+++ src/modules/dish/dish.module.ts
@@ -1,5 +1,4 @@
 import { Module } from '@nestjs/common';
-import { DishOrchestrator } from './dish.orchestrator';
 import { MongooseModule } from '@nestjs/mongoose';
 import { DishController } from './dish.controller';
 import { dishCommentModel, dishModel, dishRatingModel, userSearchQueryModel } from '../../common/definitions/mongoose-model.definitions';
@@ -7,25 +6,18 @@ import { RedisModule } from '../redis/redis.module';
 import { JwtManagerModule } from '../jwt-manager/jwt-manager.module';
 import { IngredientModule } from '../ingredient/ingredient.module';
 import { UserSearchQueryRepository } from '../../mongodb/repositories/user-search-query.repository';
-import { TranslationModule } from '../translation/translation.module';
-import { DishReadModule } from './read/dish-read.module';
-import { DishWriteModule } from './write/dish-write.module';
+import { DishApplicationModule } from './application/dish-application.module';
 
 @Module({
     imports: [
         MongooseModule.forFeature([dishModel, dishCommentModel, dishRatingModel, userSearchQueryModel]),
-        DishReadModule,
-        DishWriteModule,
+        DishApplicationModule,
         RedisModule, // FIXME: Discard?
         JwtManagerModule, // FIXME: Discard
-        IngredientModule,
-        TranslationModule // FIXME: Discard?
+        IngredientModule
     ],
     controllers: [DishController],
-    providers: [
-        DishOrchestrator,
-        UserSearchQueryRepository
-    ],
+    providers: [UserSearchQueryRepository],
     exports: []
 })
 export class DishModule {}